helpers(refactor) merged helper files into utils and updated affected files
