--- src/helpers/handle-error.js
@@ -1,3 +1,0 @@
-export const handleError = (err) => {
-  console.error(`error: ${err.message}`)
-}
