--- src/helpers/handle-response.js
@@ -1,5 +1,0 @@
-export const handleResponse = (response) => {
-  console.dir('handleResponse:')
-  console.dir(response.data)
-  return response.data
-}
