+++ src/utils/index.js
@@ -1,5 +1,15 @@
 import { date } from 'quasar'
 
+export const handleResponse = (response) => {
+  console.dir('handleResponse:')
+  console.dir(response.data)
+  return response.data
+}
+
+export const handleError = (err) => {
+  console.error(`error: ${err.message}`)
+}
+
 export const displayDateDifference = (data) => {
 	if (!data) {
 		return null;
